+++ package.json
@@ -19,7 +19,6 @@
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "^3.21.0",
     "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4"
   }
